gperftools@2.7: new version; homepage, URL, hash updates (#9010)
* gperftools@2.7: new version

* gperftools: update homepage & url

The URL in the homepage field of the gperftools package redirects to
its GitHub repo, and the URL for version 2.3 returns a 404 error. This
commit updates both fields to the most recent sensible values.

* gperftools: update commit hashes to sha256

This commit deletes the url field for all listed versions because it
is no longer needed as a result of updating the package url field. In
addition, all commit hashes are updated to SHA256.